@@ -1,23 +1,15 @@
 {
   "name": "legalis",
-  "version": "1.0.0",
+  "version": "0.0.1",
   "private": true,
   "scripts": {
     "ng": "ng",
-    "build": "npm run write:env -s && ng build",
-    "start": "npm run write:env -s && ng serve",
-    "start:heroku": "node server.js",
-    "serve:sw": "npm run build -s && npx http-server ./dist -p 4200",
+    "build": "ng build --service-worker",
+    "start": "ng serve",
+    "serve:sw": "yarn build && npx http-server ./dist -p 4200",
     "lint": "ng lint && stylelint \"src/**/*.scss\" --syntax scss",
-    "test": "npm run write:env -s && ng test",
-    "test:ci": "npm run write:env -s && npm run lint -s && ng test --configuration=ci",
-    "translations:extract": "ngx-translate-extract --input ./src --output ./src/translations/template.json --format=json --clean --sort",
-    "docs": "hads ./docs -o",
-    "write:env": "ngx-scripts env npm_package_version",
-    "prettier": "prettier --write \"./src/**/*.{ts,js,html,scss}\"",
-    "prettier:check": "prettier --list-different \"./src/**/*.{ts,js,html,scss}\"",
-    "postinstall": "npm run prettier -s && husky install",
-    "generate": "ng generate"
+    "test": "ng test",
+    "test:ci": "yarn lint && ng test --configuration=ci"
   },
   "dependencies": {
     "@angular-slider/ngx-slider": "^2.0.3",
@@ -43,7 +35,6 @@
     "angular-svg-icon": "^12.0.0",
     "bootstrap": "^5.0.2",
     "breakpoint-helper": "^1.0.3",
-    "e": "^0.2.1",
     "express": "^4.17.1",
     "file-saver": "^2.0.5",
     "lottie-web": "^5.8.1",
@@ -53,56 +44,37 @@
     "ngx-lottie": "^8.0.1",
     "ngx-pagination": "^5.1.1",
     "rxjs": "^6.6.3",
-    "tslib": "^2.2.0",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
     "@angular-builders/jest": "^12.1.0",
     "@angular-devkit/build-angular": "~12.1.1",
-    "@angular-eslint/builder": "~12.3.1",
-    "@angular-eslint/eslint-plugin": "~12.3.1",
-    "@angular-eslint/eslint-plugin-template": "~12.3.1",
-    "@angular-eslint/schematics": "~12.3.1",
-    "@angular-eslint/template-parser": "~12.3.1",
+    "@angular-eslint/builder": "12.3.1",
+    "@angular-eslint/eslint-plugin": "12.3.1",
+    "@angular-eslint/eslint-plugin-template": "12.3.1",
+    "@angular-eslint/schematics": "12.3.1",
+    "@angular-eslint/template-parser": "12.3.1",
     "@angular/cli": "~12.1.1",
     "@angular/compiler-cli": "~12.1.1",
     "@angular/language-service": "~12.1.1",
-    "@biesbjerg/ngx-translate-extract": "^7.0.3",
-    "@biesbjerg/ngx-translate-extract-marker": "^1.0.0",
     "@ngneat/until-destroy": "^8.0.4",
-    "@ngx-rocket/scripts": "^5.1.0",
     "@types/jest": "^26.0.19",
     "@types/node": "^12.19.12",
-    "@typescript-eslint/eslint-plugin": "4.28.2",
-    "@typescript-eslint/parser": "4.28.2",
-    "eslint": "^7.26.0",
+    "@typescript-eslint/eslint-plugin": "^5.9.1",
+    "@typescript-eslint/parser": "^5.9.1",
+    "eslint": "^8.6.0",
     "eslint-plugin-import": "latest",
     "eslint-plugin-jsdoc": "latest",
     "eslint-plugin-prefer-arrow": "latest",
-    "hads": "^3.0.0",
     "htmlhint": "^0.15.1",
     "https-proxy-agent": "^5.0.0",
-    "husky": "^7.0.1",
     "jest": "^27.0.6",
-    "prettier": "^2.2.1",
-    "pretty-quick": "^3.1.0",
     "stylelint": "~13.13.1",
     "stylelint-config-prettier": "^8.0.2",
     "stylelint-config-recommended-scss": "~4.3.0",
     "stylelint-config-standard": "~22.0.0",
     "stylelint-scss": "~3.20.1",
     "ts-node": "^10.1.0",
-    "typescript": "~4.3.2"
-  },
-  "prettier": {
-    "singleQuote": true,
-    "overrides": [
-      {
-        "files": "*.scss",
-        "options": {
-          "singleQuote": false
-        }
-      }
-    ]
+    "tslib": "^2.2.0"
   }
 }
