@@ -1,4 +1,4 @@
-/* tslint:disable:no-unused-variable */
+/* eslint-disable no-unused-vars, @typescript-eslint/no-unused-vars */
 
 import { TestBed, async } from '@angular/core/testing';
 import { HighlightPipe } from './highlight.pipe';
