@@ -14,6 +14,7 @@ import { LottieModule } from 'ngx-lottie';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { NgxSliderModule } from '@angular-slider/ngx-slider';
 import { PdfJsViewerModule } from 'ng2-pdfjs-viewer';
+import { GazetteItemComponent } from './components/gazette-item/gazette-item.component';
 
 export function playerFactory() {
   return import(/* webpackChunkName: 'lottie-web' */ 'lottie-web/build/player/lottie_light');
@@ -36,10 +37,11 @@ export function playerFactory() {
     LoaderComponent,
     FiltersComponent,
     NormativeItemComponent,
+    GazetteItemComponent,
     EmptyResultsComponent,
     PdfViewerComponent,
     HighlightPipe,
   ],
-  exports: [LoaderComponent, FiltersComponent, NormativeItemComponent, EmptyResultsComponent, HighlightPipe],
+  exports: [LoaderComponent, FiltersComponent, NormativeItemComponent, GazetteItemComponent, EmptyResultsComponent, HighlightPipe],
 })
 export class SharedModule {}
